refactor: 👷 fn + tests
